initial code rearrangement
