+++ qldpc/circuits/__init__.py
@@ -1,10 +1,14 @@
 from .common import (
+    MeasurementRecord,
+    QubitIDs,
     get_encoder_and_decoder,
     get_encoding_circuit,
     get_encoding_tableau,
     get_logical_tableau,
 )
-from .memory import get_memory_experiment
+from .memory import (
+    get_memory_experiment,
+)
 from .noise_model import (
     DepolarizingNoiseModel,
     NoiseModel,
@@ -14,8 +18,6 @@ from .noise_model import (
 from .syndrome_measurement import (
     EdgeColoring,
     EdgeColoringXZ,
-    MeasurementRecord,
-    QubitIDs,
     SyndromeMeasurementStrategy,
 )
 from .transversal import (
@@ -26,19 +28,20 @@ from .transversal import (
 )
 
 __all__ = [
+    "MeasurementRecord",
+    "QubitIDs",
     "get_encoder_and_decoder",
     "get_encoding_circuit",
     "get_encoding_tableau",
     "get_logical_tableau",
     "get_memory_experiment",
+    "get_memory_experiment_parts",
     "DepolarizingNoiseModel",
     "NoiseModel",
     "NoiseRule",
     "SI1000NoiseModel",
     "EdgeColoring",
     "EdgeColoringXZ",
-    "MeasurementRecord",
-    "QubitIDs",
     "SyndromeMeasurementStrategy",
     "get_transversal_automorphism_group",
     "get_transversal_circuit",
