+++ qldpc/circuits/common_test.py
@@ -25,6 +25,33 @@ from qldpc.math import op_to_string
 from qldpc.objects import Pauli
 
 
+def test_qubit_ids() -> None:
+    """Default qubit indices."""
+    code = codes.FiveQubitCode()
+    qubit_ids = circuits.QubitIDs.from_code(code)
+    data_ids, check_ids, ancilla_ids = qubit_ids
+    assert data_ids == list(range(len(code)))
+    assert check_ids == list(range(len(code), len(code) + code.num_checks))
+    assert not ancilla_ids
+
+    qubit_ids.add_ancilla(3)
+    assert qubit_ids.ancilla == [9, 10, 11]
+
+
+def test_measurement_record() -> None:
+    """Build and use a MeasurementRecord."""
+    record = circuits.MeasurementRecord()
+    record.append({0: [0, 1], 2: [2]})
+    assert record.num_measurements == 3
+    assert dict(record.items()) == record.qubit_to_measurements
+    assert record.get_target_rec(2) == stim.target_rec(-1)
+    assert record.get_target_rec(0) == stim.target_rec(-2)
+    with pytest.raises(ValueError, match="Qubit 1 not found"):
+        record.get_target_rec(1)
+    with pytest.raises(ValueError, match="Invalid measurement index"):
+        record.get_target_rec(0, 2)
+
+
 def test_restriction() -> None:
     """Raise an error for non-qubit codes."""
     code = codes.SurfaceCode(2, field=3)
